# crates/server/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Rockbox API v1",
    "version": "1.0",
    "contact": {
      "name": "Tsiry Sandratraina",
      "email": "tsiry.sndr@fluentci.io"
    },
    "summary": "Retrieve metadata from Rockbox Library, control playback.",
    "description": "",
    "x-logo": {
      "url": "https://avatars.githubusercontent.com/u/1327269?s=200&v=4"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6063"
    }
  ],
  "paths": {
    "/albums": {
      "get": {
        "summary": "Get Several Albums",
        "tags": [
          "Albums"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Album.v1",
                    "x-stoplight": {
                      "id": "g3x93ebnemzva"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-albums",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "The index of the first item to return. Default: 0 (the first item). Use with limit to get the next set of items."
          }
        ]
      }
    },
    "/albums/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get Album",
        "tags": [
          "Albums"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Album.v1"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-albums-id"
      }
    },
    "/albums/{id}/tracks": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get Album Tracks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track.v1",
                    "x-stoplight": {
                      "id": "hi439roy03b8p"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-albums-id-tracks",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "The index of the first item to return. Default: 0 (the first item). Use with limit to get the next set of items."
          }
        ],
        "tags": [
          "Albums"
        ]
      }
    },
    "/artists": {
      "get": {
        "summary": "Get Several Artists",
        "tags": [
          "Artists"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Artist.v1",
                    "x-stoplight": {
                      "id": "03rsibvu1zlxz"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-artists",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "The index of the first item to return. Default: 0 (the first item). Use with limit to get the next set of items."
          }
        ]
      }
    },
    "/artists/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get Artist",
        "tags": [
          "Artists"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Artist.v1",
                    "x-stoplight": {
                      "id": "mq2z0jvupvbjs"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-artists-id"
      }
    },
    "/tracks": {
      "get": {
        "summary": "Get Several Tracks",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track.v1",
                    "x-stoplight": {
                      "id": "1onvs6nmuxc8z"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-tracks",
        "requestBody": {
          "content": {},
          "description": ""
        },
        "x-internal": false,
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "The index of the first item to return. Default: 0 (the first item). Use with limit to get the next set of items."
          }
        ],
        "tags": [
          "Tracks"
        ]
      }
    },
    "/tracks/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get Track",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track.v1"
                },
                "examples": {
                  "Track": {
                    "value": {
                      "album": "2014 Forest Hills Drive",
                      "album_artist": "J. Cole",
                      "album_id": "cm1ibibnn00043fj7ppakq9w9",
                      "artist": "J. Cole",
                      "artist_id": "cm1ibibnn00033fj7hy019na5",
                      "bitrate": 96,
                      "composer": "Roy Hammond, Jermaine Cole, CHARLES SIMMONS",
                      "created_at": 1727296114,
                      "disc_number": 1,
                      "filesize": 2872773,
                      "frequency": 44100,
                      "genre_id": "",
                      "id": "cm1ibibnn00053fj79ghe6b2d",
                      "length": 239381,
                      "md5": "b5a0d86e156e6d02c90f647bf5cc6fc2",
                      "path": "/home/coder/Music/03 - J. Cole - Wet Dreamz(Explicit).m4a",
                      "title": "Wet Dreamz",
                      "track_number": 3,
                      "updated_at": 1727296114,
                      "year": 2014,
                      "year_string": "2014-12-09"
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-tracks-trackid",
        "tags": [
          "Tracks"
        ]
      }
    },
    "/artists/{id}/albums": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Get Artist's Albums",
        "tags": [
          "artists"
        ],
        "responses": {},
        "operationId": "get-artists-id-albums",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "The maximum number of items to return. Default: 20. Minimum: 1. Maximum: 50."
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "The index of the first item to return. Default: 0 (the first item). Use with limit to get the next set of items."
          }
        ]
      }
    },
    "/browse/tree-entries": {
      "get": {
        "summary": "Get Entries",
        "tags": [
          "Browse"
        ],
        "responses": {},
        "operationId": "get-tree_entries",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "description": "path"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "show_hidden"
          }
        ]
      },
      "parameters": []
    },
    "/player/play": {
      "parameters": [],
      "put": {
        "summary": "Play",
        "operationId": "put-player-play",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Player"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "elapsed"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset"
          }
        ]
      }
    },
...
    "/playlists": {
      "post": {
        "summary": "Create New Playlist",
        "operationId": "post-playlists",
        "responses": {
          "200": {
            "description": "OK",
            "content": {}
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "x-stoplight": {
                      "id": "74lnjkz6fpkw3"
                    }
                  },
                  "tracks": {
                    "type": "array",
                    "x-stoplight": {
                      "id": "vbmd8a6gj0bh1"
                    },
                    "items": {
                      "x-stoplight": {
                        "id": "ydqsysvo8ulzo"
                      },
                      "type": "string"
                    }
                  }
                }
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "name": "demo",
                    "tracks": [
                      "/home/coder/Music/03 - J. Cole - Wet Dreamz(Explicit).m4a"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Playlists"
        ]
      }
    },
    "/playlists/start": {
      "put": {
        "summary": "Start Playlist",
        "operationId": "put-playlists-start",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "start_index"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "elapsed"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "offset"
          }
        ]
      }
    },
    "/playlists/shuffle": {
      "put": {
        "summary": "Shuffle Playlist",
        "operationId": "put-playlists-shuffle",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "start_index"
          }
        ]
      }
    },
...
    "/artists/{id}/tracks": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [
          "Artists"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Track.v1",
                    "x-stoplight": {
                      "id": "vcpked2x8fog7"
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "get-artists-id-tracks"
      }
    }
  },
  "components": {
    "schemas": {
      "Album.v1": {
        "type": "object",
        "x-examples": {
          "Example 1": {
            "album_art": "216ccc791352fbbffc11268b984db19a.jpg",
            "artist": "J. Cole",
            "id": "cm1ibibn600013fj7d96nsaeh",
            "md5": "d55c5a429200828bc798d9f10253c1af",
            "title": "2014 Forest Hills Drive",
            "year": 2014,
            "year_string": "2014-12-09"
          }
        },
        "examples": [
          {
            "album_art": "216ccc791352fbbffc11268b984db19a.jpg",
            "artist": "J. Cole",
            "id": "cm1ibibn600013fj7d96nsaeh",
            "md5": "d55c5a429200828bc798d9f10253c1af",
            "title": "2014 Forest Hills Drive",
            "year": 2014,
            "year_string": "2014-12-09"
          }
        ],
        "title": "Album",
        "properties": {
          "album_art": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "md5": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "year_string": {
            "type": "string"
          },
          "tracks": {
            "type": "array",
            "x-stoplight": {
              "id": "7ng5gjun7ppdf"
            },
            "items": {
              "$ref": "#/components/schemas/Track.v1",
              "x-stoplight": {
                "id": "eoxt1tuziuzrf"
              }
            }
          }
        }
      },
      "Artist.v1": {
        "x-stoplight": {
          "id": "03rsibvu1zlxz"
        },
        "type": "object",
        "x-examples": {
          "Example 1": {
            "id": "cm1ibibn600003fj70egpmcbc",
            "name": "J. Cole"
          }
        },
        "examples": [
          {
            "id": "cm1ibibn600003fj70egpmcbc",
            "name": "J. Cole"
          }
        ],
        "title": "Artist",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "x-stoplight": {
              "id": "f2du13h4su3yn"
            }
          },
          "image": {
            "type": "string",
            "x-stoplight": {
              "id": "69j27eohqstp7"
            }
          },
          "albums": {
            "type": "array",
            "x-stoplight": {
              "id": "2p7oo0nysxxe0"
            },
            "items": {
              "$ref": "#/components/schemas/Album.v1",
              "x-stoplight": {
                "id": "zicm4vo4217x8"
              }
            }
          },
          "tracks": {
            "type": "array",
            "x-stoplight": {
              "id": "w0a6febmx5z3g"
            },
            "items": {
              "$ref": "#/components/schemas/Track.v1",
              "x-stoplight": {
                "id": "q4daivn3dx1gr"
              }
            }
          }
        }
      },
...
      "Track.v1": {
        "type": "object",
        "properties": {
          "album": {
            "type": "string"
          },
          "album_artist": {
            "type": "string"
          },
          "album_id": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "artist_id": {
            "type": "string"
          },
          "bitrate": {
            "type": "integer"
          },
          "composer": {
            "type": "string"
          },
          "created_at": {
            "type": "integer"
          },
          "disc_number": {
            "type": "integer"
          },
          "filesize": {
            "type": "integer"
          },
          "frequency": {
            "type": "integer"
          },
          "genre_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "length": {
            "type": "integer"
          },
          "md5": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "track_number": {
            "type": "integer"
          },
          "updated_at": {
            "type": "integer"
          },
          "year": {
            "type": "integer"
          },
          "year_string": {
            "type": "string"
          }
        },
        "x-examples": {
          "Example 1": {
            "album": "2014 Forest Hills Drive",
            "album_artist": "J. Cole",
            "album_id": "cm1ibibnn00043fj7ppakq9w9",
            "artist": "J. Cole",
            "artist_id": "cm1ibibnn00033fj7hy019na5",
            "bitrate": 96,
            "composer": "Roy Hammond, Jermaine Cole, CHARLES SIMMONS",
            "created_at": 1727296114,
            "disc_number": 1,
            "filesize": 2872773,
            "frequency": 44100,
            "genre_id": "",
            "id": "cm1ibibnn00053fj79ghe6b2d",
            "length": 239381,
            "md5": "b5a0d86e156e6d02c90f647bf5cc6fc2",
            "path": "/home/coder/Music/03 - J. Cole - Wet Dreamz(Explicit).m4a",
            "title": "Wet Dreamz",
            "track_number": 3,
            "updated_at": 1727296114,
            "year": 2014,
            "year_string": "2014-12-09"
          }
        },
        "examples": [
          {
            "album": "2014 Forest Hills Drive",
            "album_artist": "J. Cole",
            "album_id": "cm1ibibnn00043fj7ppakq9w9",
            "artist": "J. Cole",
            "artist_id": "cm1ibibnn00033fj7hy019na5",
            "bitrate": 96,
            "composer": "Roy Hammond, Jermaine Cole, CHARLES SIMMONS",
            "created_at": 1727296114,
            "disc_number": 1,
            "filesize": 2872773,
            "frequency": 44100,
            "genre_id": "",
            "id": "cm1ibibnn00053fj79ghe6b2d",
            "length": 239381,
            "md5": "b5a0d86e156e6d02c90f647bf5cc6fc2",
            "path": "/home/coder/Music/03 - J. Cole - Wet Dreamz(Explicit).m4a",
            "title": "Wet Dreamz",
            "track_number": 3,
            "updated_at": 1727296114,
            "year": 2014,
            "year_string": "2014-12-09"
          }
        ],
        "description": "",
        "title": "Track"
      }
    }
  },
...
}